BROKEN: Does not respond to path /fileName PUT if fileName does not exist

+++ routers/user-sub-routes/users-user-files-file.js
@@ -1,37 +1,44 @@
 module.exports = function(router, mongoose, bodyParser, EventEmitter, ee, User,
   File, AWS, s3, ourBucket) {
 
+    function sendResSuccess(res, data) {
+      res.json({msg: data});
+    }
+
+    function sendError404(res, data) {
+      res.status(404).json({msg: data + ' not found.'});
+    }
+
     router.route('/:user/files/:file')
     .get(function(req, res) {
       var userId = req.params.user;
       var searchedFile = req.params.file;
-
       User.findOne({username: userId}, function(err, user) {
         if (err) {
           return console.log(err);
         }
-
-        for (var i = 0; i < user.files.length; i++) {
-          (function(i) {
-            var fileId = user.files[i];
-            File.findById(fileId, function(err, file) {
-              if (file.fileName === searchedFile) {
-                var params = {Bucket: ourBucket, Key: userId + '/' + file.fileName};
-                s3.getObject(params, function(err, data) {
-                  if (err) {
-                    return console.log(err);
-                  }
-                  var foundFile = data;
-                  sendResponse(foundFile);
-                });
-              }
-            });
-          })(i);
+        if (user.files.length > 0) {
+          for (var i = 0; i < user.files.length; i++) {
+            (function(i) {
+              var fileId = user.files[i];
+              File.findById(fileId, function(err, file) {
+                if (file.fileName === searchedFile) {
+                  var params = {Bucket: ourBucket, Key: userId + '/' + file.fileName};
+                  console.log(file);
+                  s3.getSignedUrl('getObject', params, function(err, url) {
+                    if (err) return console.log(err);
+                    File.update({_id: fileId}, {$set: {"url": url}}, function(err, file) {
+                    sendResSuccess(res, url);
+                    });
+                  });
+                }
+              });
+            })(i);
+          }
+        } else {
+          sendError404(res, searchedFile);
         }
       });
-      function sendResponse(foundFile) {
-        res.json({msg: 'File contents: ' + foundFile.Body.toString()});
-      }
     })
     .put(function(req, res) {
       var userId = req.params.user;
@@ -42,7 +49,6 @@ module.exports = function(router, mongoose, bodyParser, EventEmitter, ee, User,
         if (err) {
           return console.log(err);
         }
-
         for (var i = 0; i < user.files.length; i++) {
           (function(i) {
             var fileId = user.files[i];
@@ -58,27 +64,34 @@ module.exports = function(router, mongoose, bodyParser, EventEmitter, ee, User,
                   if (err) return console.log(err);
                   file.url = url;
                 });
+
                 s3.putObject(params, function(err, data) {
                   if (err) {
                     return console.log(err);
                   }
                 });
-                file.content = newFileContents;
-                file.save();
-                sendResponse(file.fileName);
+                // file.content = newFileContents;
+                console.log(newFileContents);
+                console.log(fileId);
+                file.update({$set: {"content": newFileContents}}, function(err, data) {
+                  if (err) console.log(err);
+                });
+                User.findOne({username: userId}, function(err, user) {
+                  // user.files.push(file._id);
+                  user.update({$set: {files: file._id}}, {upsert: true});
+                  console.log(user);
+                });
+                sendResSuccess(res, file.fileName + ' updated');
               }
             });
           })(i);
         }
       });
-      function sendResponse(file) {
-        res.json({msg: file + ' was successfully saved'});
-      }
     })
     .delete(function(req, res) {
       var userId = req.params.user;
       var currFile = req.params.file;
-      
+
       var params = {
         Bucket: ourBucket,
         Delete: {
@@ -96,24 +109,27 @@ module.exports = function(router, mongoose, bodyParser, EventEmitter, ee, User,
         if (err) {
           return console.log(err);
         }
-
-        for (var i = 0; i < user.files.length; i++) {
-          (function(i) {
-            var fileId = user.files[i];
-            File.findById(fileId, function(err, file) {
-              if (file.fileName === currFile) {
-                file.remove();
-                user.update({$pull: {files: fileId}}, function(err, data){
-                  if (err) {
-                    return console.log(err);
-                  }
-                  console.log(data);
-                });
-              }
-            })
-          })(i);
+        if (user.files.length > 0) {
+          for (var i = 0; i < user.files.length; i++) {
+            (function(i) {
+              var fileId = user.files[i];
+              File.findById(fileId, function(err, file) {
+                if (file.fileName === currFile) {
+                  file.remove();
+                  user.update({$pull: {files: fileId}}, function(err, data){
+                    if (err) {
+                      return console.log(err);
+                    }
+                    console.log(data);
+                    sendResSuccess(res, data.Deleted);
+                  });
+                }
+              })
+            })(i);
+          }
+        } else {
+          sendError404(res, currFile);
         }
       });
-      res.json({msg: currFile +  ' was successfully deleted!'});
     });
 }
